pyproject.toml
[tool.poetry]
name = "aleph-vortex-app"
version = "0.0.1"
description = "Aleph.im-based app for indexing Vortex Network"
authors = ["Charaf ZELLOU <charaf.zellou@smart-chain.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
requests = "^2.31.0"
environs = "^9.5.0"
fastapi = "^0.99.1"
tartiflette-asgi = "^0.12.0"
uvicorn = "^0.22.0"
psycopg2-binary = "^2.9.6"
asyncio = "^3.4.3"
tartiflette = "^1.4.1"
asyncpg = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"